% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsmr.R
\name{bi_gsmr}
\alias{bi_gsmr}
\title{Bi-directional GSMR analysis}
\usage{
bi_gsmr(bzx, bzx_se, bzx_pval, bzy, bzy_se, bzy_pval, ldrho, snpid, heidi_outlier_flag=T, gwas_thresh=5e-8, single_snp_heidi_thresh=0.01, multi_snp_heidi_thresh=0.01, nsnps_thresh=10, ld_r2_thresh=0.05, ld_fdr_thresh=0.05, gsmr2_beta=0)
}
\arguments{
\item{bzx}{vector, SNP effects on risk factor}

\item{bzx_se}{vector, standard errors of bzx}

\item{bzx_pval}{vector, p values for bzx}

\item{bzy}{vector, SNP effects on disease}

\item{bzy_se}{vector, standard errors of bzy}

\item{bzy_pval}{vector, p values for bzy}

\item{ldrho}{LD correlation matrix of the SNPs}

\item{snpid}{genetic instruments}

\item{n_ref}{sample size of the reference sample}

\item{heidi_outlier_flag}{flag for HEIDI-outlier analysis}

\item{gwas_thresh}{threshold p-value to select  instruments from GWAS for risk factor}

\item{single_snp_heidi_thresh}{p-value threshold for single-SNP-based HEIDI-outlier analysis}

\item{multi_snp_heidi_thresh}{p-value threshold for multi-SNP-based HEIDI-outlier analysis}

\item{nsnps_thresh}{the minimum number of instruments required for the GSMR analysis (we do not recommend users to set this number smaller than 10)}

\item{ld_r2_thresh}{LD r2 threshold to remove SNPs in high LD}

\item{ld_fdr_thresh}{FDR threshold to remove the chance correlations between SNP instruments}

\item{gsmr2_beta}{GSMR2 beta version, including a new HEIDI-outlier method (used in a GSMR analysis) that is currently under development and subject to future changes, 0 - the original HEIDI-outlier method, 1 - the new HEIDI-outlier method}
}
\value{
Estimate of causative effect of risk factor on disease (forward_bxy), the corresponding standard error (forward_bxy_se), p-value (forward_bxy_pval) and SNP index (forward_index), and estimate of causative effect of disease on risk factor (reverse_bxy), the corresponding standard error (reverse_bxy_se), p-value (reverse_bxy_pval), SNP index (reverse_index), SNPs with missing values, with non-significant p-values and those in LD.
}
\description{
Bi-directional GSMR analysis is composed of a forward-GSMR analysis and a reverse-GSMR analysis that uses SNPs associated with the disease (e.g. at  < 5e-8) as the instruments to test for putative causal effect of the disease on the risk factor.
}
\examples{
data("gsmr")
gsmr_result = bi_gsmr(gsmr_data$bzx, gsmr_data$bzx_se, gsmr_data$bzx_pval, gsmr_data$bzy, gsmr_data$bzy_se, gsmr_data$bzy_pval, ldrho, gsmr_data$SNP, n_ref, T, 5e-8, 0.01, 0.01, 10, 0.05, 0.05, 0) 

}
