% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSMR-package.r
\docType{package}
\name{gsmr-package}
\alias{gsmr-package}
\title{gsmr: A tool for GSMR and HEIDI analysis}
\description{
Perform Generalized Summary-data-based Mendelian Randomization analysis (GSMR) 
and HEterogeneity In Dependent Instruments analysis to remove pleiotropic outliers (HEIDI-outlier).
}
\references{
Zhu, Z. et al. Causal associations between risk factors and common diseases inferred from GWAS summary data. Nature Communications, in press. An early verison of the manuscript is available at bioRxiv, 168674.
}
\author{
Zhihong Zhu <z.zhu1@uq.edu.au>

Zhili Zheng <zhili.zheng@uq.edu.au>

Futao Zhang <f.zhang5@uq.edu.au>

Jian Yang <jian.yang@uq.edu.au>
}
