% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsmr.R
\name{gsmr}
\alias{gsmr}
\title{Generalized Summary-data-based Mendelian Randomization analysis}
\usage{
gsmr(bzx, bzx_se, bzx_pval, bzy, bzy_se, ldrho, snpid, heidi_outlier_flag=T, gwas_thresh=5e-8, single_heidi_thresh=0.01, multi_heidi_thresh=0.01, nsnps_thresh=10, ld_r2_thresh=0.05, ld_fdr_thresh=0.05, gsmr2_beta=0)
}
\arguments{
\item{bzx}{vector, SNP effects on risk factor}

\item{bzx_se}{vector, standard errors of bzx}

\item{bzx_pval}{vector, p values for bzx}

\item{bzy}{vector, SNP effects on disease}

\item{bzy_se}{vector, standard errors of bzy}

\item{ldrho}{LD correlation matrix of the SNPs}

\item{snpid}{genetic instruments}

\item{n_ref}{sample size of the reference sample}

\item{heidi_outlier_flag}{flag for HEIDI-outlier analysis}

\item{gwas_thresh}{threshold p-value to select instruments from GWAS for risk factor}

\item{nsnps_thresh}{the minimum number of instruments required for the GSMR analysis (we do not recommend users to set this number smaller than 10)}

\item{ld_r2_thresh}{LD r2 threshold to remove SNPs in high LD}

\item{ld_fdr_thresh}{FDR threshold to remove the chance correlations between SNP instruments}

\item{gsmr2_beta}{GSMR2 beta version, including a new HEIDI-outlier method (used in a GSMR analysis) that is currently under development and subject to future changes, 0 - the original HEIDI-outlier method, 1 - the new HEIDI-outlier method}

\item{single_heidi_thresh}{p-value threshold for single-SNP-based HEIDI-outlier analysis}

\item{multi_heidi_thresh}{p-value threshold for multi-SNP-based HEIDI-outlier analysis}
}
\value{
Estimate of causative effect of risk factor on disease (bxy), the corresponding standard error (bxy_se), p-value (bxy_pval), SNP index (used_index), SNPs with missing values, with non-significant p-values and those in LD.
}
\description{
GSMR (Generalised Summary-data-based Mendelian Randomisation) is a flexible and powerful approach that utilises multiple genetic instruments to test for causal association between a risk factor and disease using summary-level data from independent genome-wide association studies.
}
\examples{
data("gsmr")
gsmr_result = gsmr(gsmr_data$bzx, gsmr_data$bzx_se, gsmr_data$bzx_pval, gsmr_data$bzy, gsmr_data$bzy_se, ldrho, gsmr_data$SNP, n_ref, T, 5e-8, 0.01, 0.01, 10, 0.1, 0.05, 0) 

}
